+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added the `higlass-zarr-datafetchers` dependency.
 - Added a demo containing a HiGlass component with a `horizontal-multivec` track using a `zarr-multivec` plugin data fetcher.
 - Add auto-sliders to layer controller.
+- Scalebar for spatial component.
 
 ### Changed
 - In the SpatialSubscriber component, no longer display counts for cells, molecules, or locations if the count value is zero.